md/raid5: Fix divide type in raid5_make_request()
0day reports a build failure on the hexagon architecture:

  ld.lld: error: undefined symbol: __hexagon_udivdi3
     referenced by raid5.c
        md/raid5.o:(raid5_make_request) in archive drivers/built-in.a
     referenced by raid5.c
        md/raid5.o:(raid5_make_request) in archive drivers/built-in.a
     did you mean: __hexagon_udivsi3
        defined in: arch/hexagon/built-in.a(lib/udivsi3.o)

This is caused by using DIV_ROUND_UP on a sector_t type.

The actual value is known to be less than 256 so a wide 64bit divide
here is not desirable. Thus cast the argument to an int to ensure it
uses a 32bit divide.

Fixes: 681fb14a7100 ("md/raid5: Pivot raid5_make_request()")
Reported-by: kernel test robot <lkp@intel.com>

+++ drivers/md/raid5.c
@@ -6091,7 +6091,7 @@ static bool raid5_make_request(struct mddev *mddev, struct bio * bi)
 	bi->bi_next = NULL;
 
 	bitmap_set(ctx.sectors_to_do, 0,
-		   DIV_ROUND_UP(ctx.last_sector - logical_sector,
+		   DIV_ROUND_UP((int)(ctx.last_sector - logical_sector),
 				RAID5_STRIPE_SECTORS(conf)));
 
 	pr_debug("raid456: %s, logical %llu to %llu\n", __func__,
